Moved Z-indexer into 'model' dir.
